test(ContactEdit): update test cases
